// src/app/modules/filing/filing-create-detail-resolution/interfaces/filing-create-detail-resolution.interface.ts
/**
 * Que es este módulo o Archivo
 *
 * Descripcion Larga
 *
 * @category     Gestion Documental
 * @package      Orfeo NG 
 * @subpackage   XXXX 
 * @author       Skina Technologies SAS (http://www.skinatech.com)
 * @license      Mixta <https://orfeolibre.org/inicio/licencia-de-orfeo-ng/>
 * @license      LICENSE.md
 * @link         http://www.orfeolibre.org
 * @since        Archivo disponible desde la version 1.0.0
 *
 * @copyright    2023 Skina Technologies SAS
 */

export interface DialogData {
  idRadicado: string;
  route: string;
  authorization: string;
}

export interface GetParamsData {
  data: {
    idResoluciones: number;
    numeroResolucion: number;
    fechaResolucion: string;
    valorResolucion: number;
  }
}